Move tests to sub-directory: turshi
